+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Version 0.0.x
 
+### 0.0.14.x (2021-Jan-20)
+- Added LogDebugMessage callback to example
+- Updated CAS BACnet Stack to version v3.26.0.0
+
+### 0.0.13.x (2021-Jan-14)
+- Added BroadcastDistributionTable to example
+
 ### 0.0.12.x (2021-Dec-08)
 
 - Added CallbackDeviceCommunicationControl to example